+++ lib/components/pieces/desktop/filter/desktop_piece_filter_modal.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:sonata/design/piece_strings.dart';
+import 'package:sonata/components/pieces/desktop/filter/piece_filter_state.dart';
 import 'package:sonata/models/piece_filter.dart';
 
 class DesktopPieceFilterModal extends StatefulWidget {
@@ -34,7 +34,7 @@ class _DesktopPieceFilterModalState extends State<DesktopPieceFilterModal> {
             style: TextStyle(fontSize: 18),
           ),
           const Divider(),
-          _StateFilter(stateNotifier),
+          PieceFilterState(stateNotifier),
         ],
       ),
       actions: [
@@ -54,49 +54,4 @@ class _DesktopPieceFilterModalState extends State<DesktopPieceFilterModal> {
         instrument: null,
         attachmentType: null,
       );
-}
-
-class _StateFilter extends StatefulWidget {
-  const _StateFilter(this.notifier);
-
-  final ValueNotifier<List<int>> notifier;
-
-  @override
-  State<_StateFilter> createState() => __StateFilterState();
-}
-
-class __StateFilterState extends State<_StateFilter> {
-  // final List<bool> isSelected = pieceStates.keys.map((e) => false).toList();
-  late List<bool> isSelected;
-
-  @override
-  void initState() {
-    super.initState();
-    isSelected = List.generate(pieceStates.length,
-        (index) => (widget.notifier.value.contains(index))).toList();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ToggleButtons(
-        // renderBorder: false,
-        // fillColor: Colors.white,
-        isSelected: List.generate(pieceStates.length,
-            (index) => (widget.notifier.value.contains(index))).toList(),
-        onPressed: (i) {
-          setState(() {
-            if (widget.notifier.value.contains(i)) {
-              widget.notifier.value.remove(i);
-            } else {
-              widget.notifier.value.add(i);
-            }
-          });
-        },
-        children: pieceStates.values
-            .map((e) => Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: Text(e),
-                ))
-            .toList());
-  }
 }
