+++ lib/components/pieces/mobile/mobile_piece_tile.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:sonata/components/pieces/mobile_piece_tag.dart';
+import 'package:sonata/components/pieces/mobile/mobile_piece_tag.dart';
 import 'package:sonata/design/piece_strings.dart';
 import 'package:sonata/models/piece.dart';
