+++ lib/pages/pieces/pieces_page/desktop/desktop_pieces_page.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 import 'package:sonata/components/desktop_navigation_drawer.dart';
-import 'package:sonata/components/pieces/desktop_piece_filter_modal.dart';
-import 'package:sonata/components/pieces/desktop_piece_row.dart';
+import 'package:sonata/components/pieces/desktop/filter/desktop_piece_filter_modal.dart';
+import 'package:sonata/components/pieces/desktop/desktop_piece_row.dart';
 import 'package:sonata/models/piece.dart';
 import 'package:sonata/models/piece_filter.dart';
 import 'package:sonata/state/global_state.dart';
